Consolidate Adyen, AdyenCard, AdyenActions bundle extensions into single file with conditional compilation using SWIFT_PACKAGE
