--- Adyen/Utilities/Non SPM Bundle Extension/CoreBundleExtension.swift
@@ -1,23 +1,0 @@
-//
-// Copyright (c) 2020 Adyen N.V.
-//
-// This file is open source and available under the MIT license. See the LICENSE file for more info.
-//
-
-import Foundation
-
-/// This is excluded from the Swift Package, since swift packages has different code to access internal resources.
-/// The Bundle extension in `BundleSPMExtension.swift` is used instead.
-extension Bundle {
-
-    /// The main bundle of the framework.
-    internal static let core: Bundle = .init(for: AdyenContext.self)
-
-    /// The bundle in which the framework's resources are located.
-    package static let coreInternalResources: Bundle = {
-        let url = core.url(forResource: "Adyen", withExtension: "bundle")
-        let bundle = url.flatMap { Bundle(url: $0) }
-        return bundle ?? core
-    }()
-
-}
